Refactor Flink Submission Service: consolidate submit logic and remove deprecated /submit-zip endpoint
